+++ levels/level2.txt
@@ -1,2 +1,2 @@
-5m;4m;5m;6m;3m;;;;;;3m;6m;1n;;1n
+5m;4m;5m;6m;3m;;;;;;3m;6m;1n;;;1n
 9
